fix(yandex-cloud): rename the MDB catalog off the banned umbrella words
main is red. The Managed Databases PR (#5828) landed a module and a class
named for "engine", which tests/quality/test_no_umbrella_names.py bans, so
test (integrations-and-misc-2) fails on every push:

  assert ['integrations/yandex_cloud/mdb_engines.py:class: Engine',
          'integrations/yandex_cloud/mdb_engines.py:module: mdb_engines'] == []

Rename both, which is what the guard's own failure message asks for:

  mdb_engines.py -> mdb_catalog.py   metadata and a discovery table, which
                                     AGENTS.md calls a *_catalog.py
  class Engine   -> ManagedDatabase

ENGINES, ENGINE_KEYS, resolve_engine and engine_choices keep their names.
The guard only reads module and class names, and "engine" is the parameter
callers actually type in both tools' schemas and read back in their output;
renaming those would drift the internals away from the public contract.
